feat: add compile-time server feature gates and fix all i18n migration errors
Add server-* Cargo features for compile-time removal of server-dependent
code in OSS builds. When server-* features are disabled, corresponding
trait implementations, struct fields, accessor methods, and consumer code
are excluded from compilation entirely, guaranteeing zero server code in
the OSS binary.

Server feature gates added:
- server-auth, server-ai, server-cloud, server-session, server-remote
- server-team, server-integration, server-block-share, server-autoupdate
- server-experiment, server-all (meta-feature)

Key changes:
- Cargo.toml: 10 server-* features + oss-default feature set; firebase,
  remote_server, warp_managed_secrets made optional
- crates/warp_features: 10 ServerXxx runtime FeatureFlag variants
- app/src/lib.rs: compound #[cfg(all(feature=X, feature="server-xxx"))]
  gates in enabled_features() + ServerXxx FeatureFlag mappings
- app/src/server/server_api: cfg-gated sub-modules, struct fields, trait
  impls (AuthClient, AIClient, ObjectClient, TeamClient, etc.), and
  ServerApiProvider accessor methods
- app/src/server/mod.rs: cfg-gated sub-modules (block, cloud_objects,
  experiments, graphql, sync_queue, telemetry, voice_transcriber)

i18n migration fixes (309 → 0 errors):
- SettingSchemaEntry description fields: &str → String (.to_string())
- LazyLock<String> statics: use .as_str() for &str, .clone() for String
- tr!() temporary values: bind to local variable before taking reference
- ToFluentValue: add i32, u64, u32, &String impls in warp_i18n
- SearchBar: &str → String/impl Into<String> for placeholder_text
- Fix block expression trailing commas (.finish(), → .finish())
- Remove dead code: unused statics, constants, imports
- Fix TERMINAL_SLOW_BOOTSTRAP naming convention
- Remove redundant .clone() on &str parameters in tr!() calls

Production build (cargo check --bin warp) compiles with 0 errors.

+++ app/src/ai/blocklist/view_util.rs
@@ -157,7 +157,7 @@ pub fn get_ai_block_overflow_menu_element_position_id(view_id: EntityId) -> Stri
 pub fn format_credits(credits: f32) -> String {
     // If the first part of the decimal is 0, we just display the whole number.
     if credits.fract() < 0.1 {
-        let whole = credits.trunc() as i32;
+        let whole = credits.trunc() as i64;
         if whole == 1 {
             crate::tr!("ai_assistant", "ai-assistant-one-credit", count = whole)
         } else {
